new plugin system
scripts are no longer wrapped in div tags in plugins and are now injected into the dom instead of removed with the div on plugin.close(). var declaration removed from object creation to allow for deleting from global scope when plugin is destroyed. data-window no longer needs to be declared when writing plugins as they are appended by plugin.load