tileset manager fixed
There was an issue where if I dropped a tileset image and then add an object selection to the queue and then drop another image and add that selection to the queue, when I went to save it, it seemed like both objects were the last image dropped. I needed a way to save each object to an offscreen canvas and then add them all as a single image to the server. This has now been fixed.