handling selection/deletion & fixed light removal/deselection
When selecting objects, it selects the item in the layers panel. when deleted with either the delete or clicking on the trash can, it will remove the object(s) from the scene, and remove the item from the layers panel and also removes any light sources and deselects the objects. This also works with parent layers. If I delete a layer that contains multiple items, they will all be removed.